+++ src/components/home.js
@@ -1,4 +1,4 @@
-function home(navigateTo) {
+function home (navigateTo) {
 
   const contenedorGeneral = document.createElement('div');
   contenedorGeneral.className = 'contenedor';
