+++ src/lib/initializeFirebase.js
@@ -10,4 +10,5 @@ export const app = initializeApp(firebaseConfig);
 const auth = getAuth(app);
 const db = getFirestore(app);
 
-export {auth,createUserWithEmailAndPassword,updateProfile,db,collection,addDoc}
+export {auth,createUserWithEmailAndPassword,updateProfile,db,collection,addDoc};
+export const saveUser = (user) => addDoc(collection(db, 'Users'), user);
